requirements-app.txt
# Databricks Apps runtime dependencies (uses Databricks Connect instead of PySpark)
databricks-connect>=14.3.0,<15
streamlit>=1.28.0
